accept new match

+++ src/main/java/com/test/springboot/demo/mycoolapp/repository/MatchRepository.java
@@ -3,6 +3,6 @@ package com.test.springboot.demo.mycoolapp.repository;
 import com.test.springboot.demo.mycoolapp.entity.Match;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatchRepository extends JpaRepository<Match, Long> {
+public interface MatchRepository extends JpaRepository<Match, Integer> {
 
 }
